requirements.txt
pillow
cairocffi
cairosvg
pangocffi
pangocairocffi
pyqrcode
pypng
